Restore the old uncached group member search logic when Member_Attr_Value is DN
Filtering on (dn=...) is invalid because DN isn't an attribute.
Therefore if DN is what we need to search on, use the member's DN as the
search base and the standard user filter.  Only if we're searching on an
attribute do we now use the standard user base and an extended user
filter.